Remove use of Mirror in Optional nil check.
The extension on KeyedEncodingContainer for OptionalEncodingWrapper
checks to make sure that the wrapped value isn't nil before
encoding. It currently constructs a Mirror reflecting the
wrapped value, and inspects that mirror's properties to
detect a nil value.

This commit replaces that code with a cast to Any and a pattern
match against Optional<Any>.none (which is nil).

This commit also adds a missing ? to the property declaration
documenting the new OptionalCoding wrapper.